// backend/src/types/sources.ts
export enum ItemCategory {
  STOCK = 'STOCK',
  CRYPTO = 'CRYPTO',
  FOREX = 'FOREX',
  COMMODITY = 'COMMODITY',
  CAR = 'CAR',
  CLOTHING = 'CLOTHING',
  COLLECTABLE = 'COLLECTABLE',
  REAL_ESTATE = 'REAL_ESTATE',
  AMAZON_PRODUCT = 'AMAZON_PRODUCT',
  ELECTRONICS = 'ELECTRONICS',
  BOOKS = 'BOOKS',
  LABOR_SERVICE = 'LABOR_SERVICE',
  FURNITURE = 'FURNITURE',
  JEWELRY = 'JEWELRY',
  SPORTS_GEAR = 'SPORTS_GEAR',
  APPLIANCES = 'APPLIANCES',
  VIDEOGAMES = 'VIDEOGAMES',
  ARTWORK = 'ARTWORK',
  GENERAL = 'GENERAL',
}

export enum DataSource {
  ALPHA_VANTAGE = 'ALPHA_VANTAGE',
  COINGECKO = 'COINGECKO',
  FIXER = 'FIXER',
  EBAY_SCRAPE = 'EBAY_SCRAPE',
  STOCKX_SCRAPE = 'STOCKX_SCRAPE',
  NEWS_API = 'NEWS_API',
  NEWS_SCRAPE = 'NEWS_SCRAPE',
  OPENAI = 'OPENAI',
}